move content manager for my own, selfish reasons
